Extracting out configuration information to its own class
